Refactor reactor tests to pytest style
also moved the test into reactor_test.py
