+++ object_database/database_test.py
@@ -30,9 +30,7 @@ from object_database.persistence import InMemoryPersistence, RedisPersistence
 from object_database.util import configureLogging, genToken
 from object_database.test_util import currentMemUsageMb
 from object_database.RedisTestHelper import RedisTestHelper
-from object_database.reactor import Reactor
 
-import object_database
 import object_database.messages as messages
 import queue
 import unittest
@@ -1658,217 +1656,6 @@ class ObjectDatabaseTests:
 
         with db1.transaction():
             self.assertEqual(len(Counter.lookupAll(k=0)), 21000)
-
-    def test_reactor_threaded(self):
-        db = self.createNewDb()
-        db.subscribeToSchema(schema)
-
-        executed = [0]
-
-        def incrementor():
-            with db.transaction():
-                for c in Counter.lookupAll():
-                    if c.x != 0:
-                        executed[0] += 1
-                        c.k += c.x
-                        c.x = 0
-
-        r1 = Reactor(db, incrementor)
-
-        with r1.running(teardown=True):
-            time.sleep(0.10)
-            self.assertEqual(executed[0], 0)
-
-            with db.transaction():
-                c = Counter(k=0, x=100)
-
-            db.waitForCondition(lambda: c.k == 100, 2.0)
-
-            with db.view():
-                self.assertEqual(c.k, 100)
-                self.assertEqual(c.x, 0)
-                self.assertEqual(executed[0], 1)
-
-            with db.transaction():
-                c.x += 2
-
-            db.waitForCondition(lambda: c.k == 102, 2.0)
-
-            with db.view():
-                self.assertEqual(c.k, 102)
-                self.assertEqual(c.x, 0)
-                self.assertEqual(executed[0], 2)
-
-    def test_reactor_with_exception(self):
-        db = self.createNewDb()
-        db.subscribeToSchema(schema)
-
-        with db.transaction():
-            c = Counter(k=0, x=0)
-
-        executed = [0]
-        thrown = [0]
-
-        def incrementor():
-            shouldThrow = False
-
-            with db.transaction():
-                executed[0] += 1
-                if c.x:
-                    thrown[0] += 1
-                    c.x = 0
-                    shouldThrow = True
-
-            assert not shouldThrow
-
-        r1 = Reactor(db, incrementor)
-
-        with r1.running():
-            for _ in range(10):
-                with db.transaction():
-                    c.x = 1
-
-                self.assertTrue(db.waitForCondition(lambda: c.x == 0, timeout=1.0))
-
-        self.assertEqual(thrown[0], 10)
-        self.assertGreater(executed[0], thrown[0])
-
-    def test_reactor_with_timestamp_lookup(self):
-        # check the semantics of 'curTimestampIsAfter'
-        t0 = time.time()
-        db = self.createNewDb()
-
-        s = Schema("schema")
-
-        @s.define
-        class Thing:
-            nextUpdate = float
-            timesUpdated = int
-
-        db.subscribeToSchema(s)
-
-        with db.transaction():
-            someThings = [Thing(nextUpdate=t0 + 0.001, timesUpdated=0) for _ in range(10)]
-
-        def incrementor():
-            with db.transaction():
-                for t in someThings:
-                    if Reactor.curTimestampIsAfter(t.nextUpdate):
-                        t.nextUpdate = t.nextUpdate + 0.01
-                        t.timesUpdated += 1
-
-        def updateCount():
-            with db.view():
-                return sum([x.timesUpdated for x in Thing.lookupAll()])
-
-        r1 = Reactor(db, incrementor)
-
-        while updateCount() < 1000 and time.time() - t0 < 2.0:
-            r1.next(timeout=1.0)
-
-        self.assertTrue(time.time() - t0 < 1.2)
-
-        r1.start()
-        self.assertTrue(
-            db.waitForCondition(
-                lambda: sum(x.timesUpdated for x in Thing.lookupAll()) >= 2000, timeout=2.0
-            )
-        )
-        self.assertTrue(time.time() - t0 < 2.2)
-        r1.stop()
-
-    def test_reactor_synchronous(self):
-        db = self.createNewDb()
-        db.subscribeToSchema(schema)
-
-        def incrementor():
-            executed = 0
-            with db.transaction():
-                for c in Counter.lookupAll():
-                    if c.x != 0:
-                        executed += 1
-                        c.k += c.x
-                        c.x = 0
-
-            return executed
-
-        r1 = Reactor(db, incrementor)
-
-        self.assertEqual(r1.next(timeout=0.01), 0)
-        self.assertIs(r1.next(timeout=0.01), object_database.reactor.Timeout)
-
-        with db.transaction():
-            c = Counter(k=0, x=100)
-
-        self.assertEqual(r1.next(timeout=0.01), 1)
-        self.assertEqual(r1.next(timeout=0.01), 0)
-        self.assertIs(r1.next(timeout=0.01), object_database.reactor.Timeout)
-
-        with db.view():
-            self.assertEqual(c.k, 100)
-            self.assertEqual(c.x, 0)
-
-        with db.transaction():
-            c.x += 2
-
-        self.assertEqual(r1.next(timeout=0.01), 1)
-        self.assertEqual(r1.next(timeout=0.01), 0)
-        self.assertIs(r1.next(timeout=0.01), object_database.reactor.Timeout)
-
-        with db.view():
-            self.assertEqual(c.k, 102)
-            self.assertEqual(c.x, 0)
-
-        r1.teardown()
-
-    def test_reactor_block_until_true(self):
-        db = self.createNewDb()
-        db.subscribeToSchema(schema)
-
-        checkCount = [0]
-        incrementCount = [0]
-
-        def incrementor():
-            incrementCount[0] += 1
-            # move '1' from any nonzero 'x' to 'k'
-            with db.transaction():
-                for c in Counter.lookupAll():
-                    if c.x > 0:
-                        c.k += 1
-                        c.x -= 1
-
-            time.sleep(0.001)
-
-        def checker():
-            # Check if any counters exist with x > 0
-            checkCount[0] += 1
-            with db.view():
-                for c in Counter.lookupAll():
-                    if c.x > 0:
-                        return False
-            return True
-
-        incrementor = Reactor(db, incrementor)
-        checker = Reactor(db, checker)
-        incrementor.start()
-
-        with db.transaction():
-            c = Counter(k=0, x=10)
-
-        self.assertFalse(checker.blockUntilTrue(timeout=0.00001))
-        self.assertTrue(checker.blockUntilTrue(timeout=1.0))
-
-        self.assertTrue(incrementCount[0] >= 10)
-        self.assertTrue(checkCount[0] >= 5)
-
-        with db.transaction():
-            c.x += 5
-
-        self.assertFalse(checker.blockUntilTrue(timeout=0.00001))
-        self.assertTrue(checker.blockUntilTrue(timeout=1.0))
-
-        self.assertTrue(incrementCount[0] >= 15)
-        self.assertTrue(checkCount[0] >= 8)
 
     def test_adding_while_subscribing(self, shouldSubscribeToIndex=False):
         pfactor = self.PERFORMANCE_FACTOR
